Add core ZeroOmega rule list config file
Initial upload of the core configuration file for ZeroOmega/SwitchyOmega.

This .sorl file contains the list of Iranian hosted domains (based on the Iran Hosted Domains project) structured for automatic traffic splitting. The configuration ensures domestic traffic is routed Direct while all other traffic is routed through the proxy profile defined in the extension.

This file is required to follow the setup guide provided in the README.md.